Created a resetState function in CustomMealAddModal. Also moved constants.ts into static, since they are just static data.
