+++ src/components/containers/table/TableCell.tsx
@@ -1,4 +1,4 @@
-import { IMPORTANCE_CLASSES } from '../../../lib/constants';
+import { IMPORTANCE_CLASSES } from '../../../static/constants';
 import {
   ImportanceIndex,
   newImportanceIndex
