+++ src/static/constants.ts
